Handle start/terminate requests in the order that were issued.
Right now is not possible to start and stop a child under the same
sync window, because the operations are not done in the same order
as requested by start/terminate child api calls.
For example, if the client issues a stop and then a start of the same
child, the child is started and then immediately terminated because
pogo performs all start operations and then all stop operations.

This commit introduce local ordering of operations:
instead of doing them separately, operations are assigned a sequence
number and on the sync interval are executed in the called order,
which ensures that pogo does what the client expects.