Handle local supervisor crashes.
The internal supervisor may indeed crash, for example when max restart
intensity is reached, which may be a transitory error.
When this happens, the entire pogo tree is restarted, loosing
all dynamic children added meanwhile.
This patch monitors the internal supervisor and restart it when dies,
without crashing the pogo GenServer. In addition, since pogo does
not re-start processes that it thinks they're already supervised,
we're ensuring to restart those processes that are supervised but not
running since the internal sup died.